+++ api/app/Http/Resources/TicketResource.php
@@ -20,6 +20,7 @@ class TicketResource extends JsonResource
             'citation_number' => $this->citation_number,
             'status' => $this->status,
             'violator' => new ViolatorResource($this->violator),
+            'deleted_at' => optional($this->deleted_at)->format('F j, Y'),
             'created_at' => $this->created_at->format('F j, Y'),
         ];
     }
